Move device info to Places entities
