+++ custom_components/places/entity.py
@@ -8,6 +8,7 @@ from typing import TYPE_CHECKING
 from homeassistant.components.sensor import SensorDeviceClass, SensorEntity, SensorEntityDescription
 from homeassistant.const import UnitOfLength
 from homeassistant.core import callback
+from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.typing import StateType
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
@@ -41,6 +42,8 @@ from .const import (
     ATTR_STREET,
     ATTR_STREET_NUMBER,
     ATTR_STREET_REF,
+    CONF_NAME,
+    DOMAIN,
 )
 
 if TYPE_CHECKING:
@@ -69,7 +72,25 @@ class PlacesEntity(CoordinatorEntity["PlacesUpdateCoordinator"]):
             if unique_suffix is None
             else f"{coordinator.config_entry.entry_id}_{unique_suffix}"
         )
-        self._attr_device_info = coordinator.device_info
+        self._attr_device_info = self.device_info
+
+    @property
+    def device_info(self) -> DeviceInfo:
+        """Return the shared HA Device metadata for this config entry."""
+        current_name = self.coordinator.get_attr_safe_str(CONF_NAME)
+        if not current_name:
+            current_name = str(
+                self.coordinator.config_entry.data.get(
+                    CONF_NAME,
+                    self.coordinator.config_entry.entry_id,
+                )
+            )
+        return DeviceInfo(
+            identifiers={(DOMAIN, self.coordinator.config_entry.entry_id)},
+            name=current_name,
+            manufacturer="Places",
+            model="OpenStreetMap reverse geocode",
+        )
 
 
 class PlacesSensorEntity(PlacesEntity, SensorEntity):
@@ -88,7 +109,7 @@ class PlacesSensorEntity(PlacesEntity, SensorEntity):
     @callback
     def _handle_coordinator_update(self) -> None:
         """Refresh cached entity attributes from coordinator data and write state."""
-        self._attr_device_info = self.coordinator.device_info
+        self._attr_device_info = self.device_info
         self._update_from_coordinator()
         self.async_write_ha_state()
